Chatbot_RAG/requirements.txt
streamlit
cohere
faiss-cpu
pypdf
python-dotenv
numpy
